backend/server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.4.0
attrs==23.2.0
Authlib==1.3.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
ecdsa==0.19.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.6.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-anthropic==0.1.20
langchain-core==0.2.18
langchain-text-splitters==0.2.2
langsmith==0.1.85
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
termcolor==2.4.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
xlrd==2.0.1
yarl==1.9.4
